example1/requirements.txt
google-cloud-translate==3.6.1